+++ src/main/java/hr/avrbanac/inanis/textures/ModelTexture.java
@@ -1,0 +1,13 @@
+package hr.avrbanac.inanis.textures;
+
+public class ModelTexture {
+    private int textureId;
+
+    public ModelTexture(int textureId) {
+        this.textureId = textureId;
+    }
+
+    public int getTextureId() {
+        return textureId;
+    }
+}