+++ CHANGELOG.md
@@ -7,6 +7,49 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## Unreleased
 - Nothing here
 
+## v0.0.7 - 2020-02-25 @avrbanac
+
+### Added
+- render must bind texture
+- render method must also activate texture
+- render method now enables and disables vertAttribArray 1 also
+- binding new VBO to VAO in ShaderProgram
+- another VBO with texture coords to model VAO
+- added TexturedModel class to represent model with texture (pojo)
+- textures package with ModelTexture (pojo)
+- Inanis static exit method (1 entry point, 1 exit point)
+- Inanis error code enumeration
+- new configuration points to global config class (InanisConfig)
+- loader now has loadTexture method
+- texture to the rendered quad
+
+### Fixed
+- bindAttribute() in AbstractShaderProgram needs to be called before link program
+- switching to camelCase variable names for shaders
+- fragment shader now uses texture() method
+- fragment shader now uses uniform sampler2D
+- fragment shader takes passed texture coordinates
+- vertex shader no longer outputs color but 2d vec (pass_texture_coords)
+- Renderer now renders TexturedModel instead of RawModel
+
+## v0.0.6 - 2020-02-24 @avrbanac
+
+### Added
+- abstract shader program and static implementation
+- basic vertex and fragment shaders for OpenGL ver 400 core (resources)
+
+## v0.0.5 - 2020-02-19 @avrbanac
+
+### Added
+- loader now has storeDataInIntBuffer method
+- implementing index buffer idea
+
+### Fixed
+- TestQuadModel now needs to provide Vertex data and Index buffer
+- Render#render now uses different method
+- Loader#loadToVAO return size of indices.length instead of positions.length/3
+- Loader#loadToVAO now takes int[] indices argument also
+
 ## v0.0.4 - 2020-02-18 @avrbanac
 
 ### Added